+++ app/public/javascript/menu_principal/menu_principal_16a_16a.js
@@ -145,18 +145,18 @@ var m = parseFloat(0.285);
 var n = parseFloat(0.356);
 
 //RDB2 = RDB1 * ESPES
-var ESPES = parseInt(form06.espessura_d.value);
-var RDB1 = parseInt(form03.raioe.value);
+var ESPES = parseFloat(form06.espessura_d.value);
+var RDB1 = parseFloat(form03.raioe.value);
 form07.raiod.value = (RDB1 * ESPES).toFixed(2) || 0.00;
-var RDB2 = parseInt(form07.raiod.value);
+var RDB2 = parseFloat(form07.raiod.value);
 
 //RN = RDB2 + ESPES / 2
 form08.raio_neutro.value = (RDB2 + ESPES / 2).toFixed(2) || 0.00;
-var RN = parseInt(form08.raio_neutro.value);
+var RN = parseFloat(form08.raio_neutro.value);
 
 //ALTDB = RDB2 + ESPES
 form09.alturadobra.value = (RDB2 + ESPES).toFixed(2) || 0.00;
-var ALTDB = parseInt(form09.alturadobra.value);
+var ALTDB = parseFloat(form09.alturadobra.value);
 
 
 //U = RN * 1.57
@@ -165,23 +165,23 @@ form10.ud.value = ((form08.raio_neutro.value * 10 * 1.57) / 10).toFixed(2) || 0.
 var U = parseFloat(form10.ud.value);
 
 //A2 = A1 - 2 * ALTDB
-var A1 = parseInt(form04.dtam_A.value);
+var A1 = parseFloat(form04.dtam_A.value);
 form09.dtam_a.value = (A1 - 2 * ALTDB).toFixed(2) || 0.00;
 var A2 = form09.dtam_a.value;
 
 
 //B2 = B1 - ALTDB
-var B1 = parseInt(form05.dtam_B.value);
+var B1 = parseFloat(form05.dtam_B.value);
 form09.dtam_b.value = B1 - ALTDB;
-var B2 = parseInt(form09.dtam_b.value);
+var B2 = parseFloat(form09.dtam_b.value);
 
 //AB = A1 - ESPES
 form05.dtam_a1.value = (A1 - ESPES).toFixed(2) || 0.00;
-var AB = parseInt(form05.dtam_a1.value);
+var AB = parseFloat(form05.dtam_a1.value);
 
 //BB = B1 - ESPES / 2
 form06.dtam_bbarra.value = (B1 - ESPES / 2).toFixed(2) || 0.00;
-var BB = parseInt(form06.dtam_bbarra.value);
+var BB = parseFloat(form06.dtam_bbarra.value);
 
 //CGX = BB
 form14.cgx.value = BB;
