+++ app/public/javascript/menu_principal/menu_principal_26_26.js
@@ -138,18 +138,18 @@ var l = parseFloat(3.1416);
 var pi = parseFloat(3.14)
 
 //RDB2 = RDB1 * ESPES
-var ESPES = parseInt(form06.espessura_d.value);
-var RDB1 = parseInt(form03.raioe.value);
+var ESPES = parseFloat(form06.espessura_d.value);
+var RDB1 = parseFloat(form03.raioe.value);
 form07.raiod.value = (RDB1 * ESPES).toFixed(2) || 0.00;
-var RDB2 = parseInt(form07.raiod.value);
+var RDB2 = parseFloat(form07.raiod.value);
 
 //RN = RDB2 + ESPES / 2
 form08.raio_neutro.value = (RDB2 + ESPES / 2).toFixed(2) || 0.00;
-var RN = parseInt(form08.raio_neutro.value);
+var RN = parseFloat(form08.raio_neutro.value);
 
 //ALTDB = RDB2 + ESPES
 form09.alturadobra.value = (RDB2 + ESPES).toFixed(2) || 0.00;
-var ALTDB = parseInt(form09.alturadobra.value);
+var ALTDB = parseFloat(form09.alturadobra.value);
 
 //U = RN * 1.57
 var U1 = parseFloat(1.57);
@@ -157,25 +157,25 @@ form10.ud.value = ((form08.raio_neutro.value * 10 * 1.57) / 10).toFixed(2) || 0.
 var U = parseFloat(form10.ud.value);
 
 //A2 = A1 - 2 * ALTDB
-var A1 = parseInt(form04.dtam_A.value);
+var A1 = parseFloat(form04.dtam_A.value);
 form09.dtam_a.value = (A1 - 2 * ALTDB).toFixed(2) || 0.00;
-var A2 = parseInt(form09.dtam_a.value);
+var A2 = parseFloat(form09.dtam_a.value);
 
 
 //B2 = B1 - ALTDB
-var B1 = parseInt(form05.dtam_B.value);
+var B1 = parseFloat(form05.dtam_B.value);
 form09.dtam_b.value = (B1 - ALTDB).toFixed(2) || 0.00;
-var B2 = parseInt(form09.dtam_b.value);
+var B2 = parseFloat(form09.dtam_b.value);
 var C1 = form03.etam_C.value
 
 
 //AB = A1 - ESPES
 form05.dtam_a1.value = (A1 - ESPES).toFixed(2) || 0.00;
-var AB = parseInt(form05.dtam_a1.value);
+var AB = parseFloat(form05.dtam_a1.value);
 
 //BB = B1 - ESPES / 2
 form06.dtam_bbarra.value = (B1 - ESPES / 2).toFixed(2) || 0.00;
-var BB = parseInt(form06.dtam_bbarra.value);
+var BB = parseFloat(form06.dtam_bbarra.value);
 
 //ALFA1 = 180 - alfa
 ALFA1 = 180 - alfa;
@@ -255,7 +255,7 @@ var XB = parseFloat(form07.dtam_xx.value);
 
 //YB = AB / 2
 form09.dtam_y.value = AB / 2;
-var YB = parseInt(form09.dtam_y.value);
+var YB = parseFloat(form09.dtam_y.value);
 
 //I1 = (((ALFA1 * 3.1416 + SENO1 * COS1 * 180) / 360) - (SENO1 * SENO1 * 180 / (ALFA1 * 3.1416))) * RN ^ 3
 I1 = (((ALFA1 * l + SENO1 * COS1 * 180) / 360) - (SENO1 * SENO1 * 180 / (ALFA1 * l))) * Math.pow(RN,3);
@@ -316,7 +316,7 @@ var CGX = parseFloat(form14.cgx.value);
 
 //CGY = YB + ESPES / 2
 form15.cgy.value = YB + ESPES / 2;
-var CGY = parseInt(form15.cgy.value);
+var CGY = parseFloat(form15.cgy.value);
 
 //X = SENO1 * ((BB - XB) + YB * COS1 / SENO1)
 X = SENO1 * ((BB - XB) + YB * COS1 / SENO1);
